+++ CHANGELOG.md
@@ -92,6 +92,7 @@
 ### Feature
 
 - /contents (folder content view) is sortable @ksuess
+- Updated the (folder content icons to Pastanaga Icons) @iFlameing
 
 ### Bugfix